Converted dday to a function

+++ src/Functions/leapy.f90
@@ -1,13 +1,13 @@
 function leapy(yr)
   implicit none
-  logical              :: leapy
-  integer, intent(in)  :: yr
+  logical           :: leapy
+  real, intent(in)  :: yr
 
-  if (mod(yr, 4) .ne. 0) then
+  if (mod(yr, 4.0) .ne. 0) then
      leapy = .false.
-  elseif(mod(yr, 25) .ne. 0) then
+  elseif(mod(yr, 25.0) .ne. 0) then
      leapy = .true.
-  elseif(mod(yr, 16) .ne. 0) then
+  elseif(mod(yr, 16.0) .ne. 0) then
      leapy = .false.
   else
      leapy = .true.
